pyproject.toml
[project]
name = "bom-defused"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "argparse>=1.4.0",
    "detect-secrets>=1.5.0",
    "pillow>=12.0.0",
    "pre-commit>=4.5.0",
    "pre-commit-hooks>=6.0.0",
    "requests>=2.32.5",
]
